Create templatePreviewMixin.js
