+++ resources/js/processes/screen-templates/components/PublicTemplatesListing.vue
@@ -16,8 +16,11 @@
       <filter-table
         :headers="fields"
         :data="data"
+        :selected-row="selectedRow"
         table-name="public-screen-templates"
         style="height: calc(100vh - 355px)"
+        @table-row-mouseover="handleRowMouseover"
+        @table-row-mouseleave="handleRowMouseleave"
       >
         <!-- Slot Table Header filter Button -->
         <template
@@ -27,7 +30,7 @@
           <div
             v-if="column.sortable"
             :key="index"
-            @click="handleEllipsisClick(column)"
+            style="display: inline-block"
           >
             <i
               :class="[
@@ -88,17 +91,6 @@
                     {{ row[header.field] }}
                   </b-tooltip>
                 </template>
-                <template v-if="header.field === 'actions'">
-                  <ellipsis-menu
-                    class="public-template-table"
-                    :actions="publicTemplateActions"
-                    :permission="permission"
-                    :data="row"
-                    :divider="true"
-                    :screen-template="true"
-                    @navigate="onTemplateNavigate"
-                  />
-                </template>
                 <template v-if="header.field !== 'name'">
                   <div :style="{ maxWidth: header.width + 'px' }">
                     {{ getNestedPropertyValue(row, header) }}
@@ -109,6 +101,39 @@
           </td>
         </template>
       </filter-table>
+      <screen-templates-tooltip
+        v-show="isTooltipVisible"
+        :position="rowPosition"
+      >
+        <template #screen-templates-tooltip-body>
+          <div
+            @mouseover="clearHideTimer"
+            @mouseleave="hideTooltip"
+          >
+            <slot
+              name="tooltip"
+              :tooltipRowData="tooltipRowData"
+              :previewTemplate="previewTemplate"
+            >
+              <span>
+                <i
+                  class="fa fa-eye py-2"
+                  @click="previewTemplate(tooltipRowData)"
+                />
+              </span>
+              <ellipsis-menu
+                class="public-template-table"
+                :actions="publicTemplateActions"
+                :data="tooltipRowData"
+                :divider="false"
+                :permission="permission"
+                :screen-template="true"
+                @navigate="onTemplateNavigate"
+              />
+            </slot>
+          </div>
+        </template>
+      </screen-templates-tooltip>
       <pagination-table
         :meta="data.meta"
         data-cy="public-templates-pagination"
@@ -124,6 +149,11 @@
         @vuetable-pagination:change-page="onPageChange"
       />
     </div>
+    <template-preview-container
+      ref="preview"
+      :selected-template="selectedTemplate"
+      @mark-selected-row="markSelectedRow"
+    />
   </div>
 </template>
 
@@ -133,15 +163,23 @@ import datatableMixin from "../../../components/common/mixins/datatable";
 import dataLoadingMixin from "../../../components/common/mixins/apiDataLoading";
 import ellipsisMenuMixin from "../../../components/shared/ellipsisMenuActions";
 import EllipsisMenu from "../../../components/shared/EllipsisMenu.vue";
+import TemplatePreviewContainer from "./TemplatePreviewContainer.vue";
+import ScreenTemplatesTooltip from "./ScreenTemplatesTooltip.vue";
 import FilterTableBodyMixin from "../../../components/shared/FilterTableBodyMixin";
 import paginationTable from "../../../components/shared/PaginationTable.vue";
 import fieldsMixin from "../mixins/fieldsMixin";
 import navigationMixin from "../mixins/navigationMixin";
+import templatePreviewMixin from "../mixins/templatePreviewMixin";
 
 const uniqIdsMixin = createUniqIdsMixin();
 
 export default {
-  components: { EllipsisMenu, paginationTable },
+  components: {
+    EllipsisMenu,
+    paginationTable,
+    TemplatePreviewContainer,
+    ScreenTemplatesTooltip,
+  },
   mixins: [
     datatableMixin,
     dataLoadingMixin,
@@ -150,6 +188,7 @@ export default {
     uniqIdsMixin,
     fieldsMixin,
     navigationMixin,
+    templatePreviewMixin,
   ],
   props: {
     permission: {
@@ -180,6 +219,13 @@ export default {
         },
       ],
       fields: [],
+      isTooltipVisible: false,
+      rowPosition: {},
+      tooltipRowData: {},
+      hideTimer: null,
+      selectedRow: 0,
+      showTemplatePreview: false,
+      selectedTemplate: null,
     };
   },
   created() {
